directory: Remove an `is_enabled` check
Summary: In the next diff, I'm going to slightly change how we enable/disable this. The result of that is going to be that this value can change out from under you. This code would be wrong if we did that - make it correct in the case where the value changes

Reviewed By: scottcao

Differential Revision: D83889173

fbshipit-source-id: 664a1e68e3eb1f7a90830b099002fadaef8479dc