directory: Accept owned path
Summary:
Instead of re-allocating a copy, accept an owned path from the caller in `LazyDirectoryBuilder::insert`. In many cases the caller already has one available

I'll note that this is not necessarily perf-neutral in the non-faster-directories case. I'll bench that case at the end of the stack to make sure we didn't introduce regressions there

Reviewed By: IanChilds

Differential Revision: D83725665

fbshipit-source-id: 9807797094f0f396d9b9e427f14a4bc2616f8b76