directory: Optimize directory merging order, part 2
Summary:
In the `LazyDirectoryBuilder`, adjust the directory merging order so that we merge smaller things into larger things.

The canonical example for why this should matter is like this: Consider some binary that depends on lib1, lib2, and zippy, with zippy also depending on lib1 and lib2. If you merge directories in that order, you end up doing a whole lot of work to merge the structures of lib1 and lib2; on the other hand, if you merge them both into zippy, since they're both a subset of zippy, we'll still need to verify that they're a subset but we won't invalidate any of the already-fingerprinted directories.

Reviewed By: IanChilds

Differential Revision: D83725661

fbshipit-source-id: 4235fa160f2e430ab035e8e10fb5190347a4453e