always run the forkserver in its own scope & slice
Summary:
Note: my end goal with this stack is to be able to log peak local action(s) memory use.

The cgroup hierarchy when local action control is enabled vs .not is structurally
very different. Here is a paste of running one action: P1973716277.

When enable_action_cgroup_pool is not enabled, there is no slice containing the
forkserver and its children, which makes it tricky to actually find the
appropriate cgroup if we want to be able to monitor the memory use of local
actions.

This diff updates this so that we just always give the forkserver its own
cgroup, that seems to not hurt anything and makes this structure a bit more
predictable?

Reviewed By: IanChilds

Differential Revision: D83744109

fbshipit-source-id: 8b552b5741dbef02acc885d271d3ddfad0b755ac