Avoid expensive copies of stdout running CT tests
Summary:
# Context
When a common-test runs, we collect the stdout of each callback and then do some manipulations on it before displaying it to the user.

# Problem
At the moment we concatenate the collected stdout, that can be large, even though we expect them to be chardata.

# This diff
Use chardata for concatenation instead

Reviewed By: michalmuskala

Differential Revision: D83932561

fbshipit-source-id: 4bc33770a0220726338b78d61c773b73a68e1786