Don't quote big ints in JSON
Summary:
Context:
https://fb.workplace.com/groups/buck2users/posts/4075951909327826

This relies on serdes 'arbitrary_precision' to handle these large ints.

Note this is definitely backwards-breaking, but given the old behavior was effectively non-deterministic (as a macro author) perhaps it's fine?

Relatively few callers in fbsource, if we need to audit them: https://fburl.com/code/z73jlcf0

Reviewed By: cjhopman

Differential Revision: D82017439

fbshipit-source-id: a47827adb33b6a729981875faaf9bc76ed80158f