Expose test target references for automatic TestableReference generation in Xcode schemes
Summary:
Original task:  T237091050

## Context
Assume you're on Xcode and that we have a target `Foo` with a corresponding test target `FooTests`. Currently, to run `FooTests`, you must manually switch to the `FooTests` target and use `cmd+U`. Ideally, if `FooTests` is set as the test target for `Foo` in the Buck build file, running `cmd+U` from the `Foo` target should execute all associated test targets automatically.

## Investigation

mavitalemeta and I investigated this today and we realized that we can add `TestableReference`s to the `Foo` scheme, such as:
{F1982457562}

What that means is, if we have a map from target `Foo` to its corresponding test target `FooTest`, we can simply add the `TestableReference`, such as:
```
...
<Testables>
         <TestableReference
            skipped = "NO">
            <BuildableReference
               BuildableIdentifier = "primary"
               BlueprintIdentifier = "29EB31DA829DEAF173B61C42"
               BuildableName = "TestFoo.xctest"
               BlueprintName = "TestFoo"
               ReferencedContainer = "container:fbobjc/Apps/Internal/FocusPlayground/FocusPlayground.xcodeproj">
            </BuildableReference>
         </TestableReference>
      </Testables>
...
```

# This diff

This diff:
- Reads info from `targets_output.json` file and exposes method `get_test_target_refs` in the current target, which generates test target references and can be passed forward so we can create `TestableReference`s
- Updates `xcscheme` fixtures:
  - `buck2 test fbsource//xplat/buck2/xcode_project:project_test -- --env UPDATE_FIXTURES=xplat/buck2/xcode_project/project_test_fixtures --run-disabled`

Differential Revision: D83870106

fbshipit-source-id: 9675640d7be8c3a7ff64c2358d3d611a33bfa85c