Fix standalone (non-tpx) test-running
Summary:
# Context
Test running starts in the `test_binary` module. It supports two modes, using TPX (internal to Meta) or standalone.

# Problem
The standalone mechanism got recently broken, as it started to pass test names as binaries where currently strings are expected.

This would have been easily spotted by eqwalizer, however this module was not eqwalized.

# This diff
Now that the rest of the utility modules are eqwalized, we eqwalize this module as well , which in the process confirmed the type-error in `listing_to_testnames()` that we now fix.

Reviewed By: michalmuskala

Differential Revision: D83932553

fbshipit-source-id: 2b4ecb350b07bfd3eebbdeb40d5817624359406c