Eqwalize list_test
Summary:
# Context
`list_test` deals with generating a list of fully-qualified test-names for each test in the `all()` definition.

# Problem
The module wasn't eqwalized. It needs to copy types from `ct_suite`, since most are not exported, and some of these types are wrong in `ct_suite` (in particular, `ct_group_def()`, as properties are optional).

# This diff
We eqwalize the module and in the process delete some function clauses that were unreachable

Reviewed By: michalmuskala

Differential Revision: D83932563

fbshipit-source-id: 1f7a17ea4b2e0ee7d2c94f880db97971a6ed45cb