log available parallelism
Summary:
Like it says in the title. I want to use this to evaluate the quality of our
predictions for memory use and since this is already being neatly packaged it's
just a matter of sending it up for logging. It also seems somewhat generally
useful anyways.

(note: AI wrote this, I reviewed it)

Reviewed By: IanChilds

Differential Revision: D84063806

fbshipit-source-id: 80b9504fb5f85a989d4d853e492f3fa3b36f6043