Update llm prompt to look at just buck2/app instead of fbcode/buck2/app
Summary: I added this llm prompt so that Devmate can automatically run `arc rust-check` on buck2/app changes but that's not always happening in practice. One hypothesis I have is that to use Devmate efficiently, I need to root my Vscode workspace at fbsource/fbcode/buck2, in which case Devmate may not be ware that buck2/app changes are part of fbcode/buck2/app. This diff makes a slight tweak to the llm prompt file just in case.

Reviewed By: IanChilds

Differential Revision: D84074281

fbshipit-source-id: 966b369d9a991a89356914631a6aba9360843bc8