Handle edge case where the path is absolute
Summary:
Context: https://fb.workplace.com/groups/apple.workflows.adhoc/permalink/1338696261244940/

Something changed in stella and /tmp/lto.tmp is showing up as inputs to the scrubber. This was leading to an infinite loop in the while where we were getting "/" as the path and we kept calling os.path.split on it.

Updating the break clause to avolid the infinite loop.

Reviewed By: wi11

Differential Revision: D84260835

fbshipit-source-id: 30a0ba32b17266dcfa11fefbd8625438c483ceb0