Add incremental action test that reproduces soft error
Summary:
When we have a fresh daemon, and we don't have an incremental output on disk, then attempting to materialize the path will fail, since the materializer doesn't have any knowledge of it (declarations to the materializer are not persisted across daemon restarts).

This is causing T239354424, which is what this test reproduces.

Reviewed By: Nero5023

Differential Revision: D84357183

fbshipit-source-id: 4ed62e21785f63edf3fcdc05f6eea60f850f26da