semanticdb subtarget
Summary:
High-level goal: Index java codebases with a BXL script

The point of this diff is to be able to use the semanticdb_javac plugin via a subtarget, as in
```
buck build $target[semanticdb] --show-simple-output
```

The way it works is similar to the [nullsafe] subtarget. In particular, the Dependency for the plugin is acquired through the JavaToolchain.

For now you have to pass the target for the semanticdb_javac plugin via `-c semanticdb_javac.plugin`, but I'm removing that in a later diff.

Clean up of the old way to invoke will be done in a later diff once we're not using that old way.

Reviewed By: IanChilds

Differential Revision: D84186609

fbshipit-source-id: 306a5e59901b70808c134d76c31ec681b3c2a5a1