Adjust dylibs extension safe check for quick tbd generation purposes
Summary:
`ResolvedStringWithMacros` objects can't be reliably compared to string literals at analysis time, and this made the existing is_extension_safe logic incorrect. The new logic also isn't 100% correct, but it should be sufficient to run an experiment to at least determine the impact of the feature.

If it proves impactful, we can thread through a separate is_extension_safe attribute, or discover the flag in a wrapper script or within the tbd generation tool itself.

Reviewed By: rmaz

Differential Revision: D84541452

fbshipit-source-id: 3d66aae1d7851c5f8c2070325d2595d9e66a2d06