Add support to set category of StarlarkActionSubError
Summary:
This diff adds support to set `category` of `StarlarkActionSubError`

This diff use `RefCell` to warp `category` to make it mutable. Since `RefCell` does not implements Clone, Ord,  PartialOrd, we just remove derive for Clone, Ord, PartialOrd which are useless for our use case.

---------
Why not implement all `set_attribute` for all other attributes?
I don't see any use cases for them right now, so to keep it simple just implement for `category`. We can implement them later if we need them.
---------

The use case for this is when we parsed an error using errorformat, we want to base on the structured error to update the category. This diff provides such ability.

One example is in the next diff.

Reviewed By: Will-MingLun-Li

Differential Revision: D84553084

fbshipit-source-id: f8e51aed015f4eb007d58c8fb387a6b708bddf6e