Expand the omnibus search space to consider native deps with SharedLibraryInfo
Summary:
**Updated description** (from Athma)

When using omnibus linking with python wheel, we currently consider extensions from the manifest associated with `libraries` and `libraries_query` attributes as potential roots for omnibus. This will miss native deps that are dlopen'ed from Python via calls like `torch.ops.load_library` leading to "Could not load this library" errors at runtime.

While the original approach in this diff passed in an explicit `omnibus_link_roots_query` to address the issue by finding additional roots, the latest version infers them from direct / first level native deps (of Python libraries) with `SharedLibraryInfo` instead, which eliminates the need for an additional attribute.

 ---

**Original Description** (from Jeremy)

There doesn't appear to be a very good way anymore for omnibus linking to determine "this library needs to be created as a *.so file on disk, with XYZ symbols in it". Non-customized CPython requires that `import _some.C.Extension` be in a module path like `_some/C/Extension.so` (or something).

I looked at just using everything with a `DlopenableLibraryInfo` provider, but that seemed to be a LOT of link roots. Just doing `cxx_python_extension` isn't right either, because some `cpp_libraries` are things that you want to call `torch.load_extension()` on.

For now, I opted on an explicit list provided by a new `roots_query`.

Reviewed By: christycylee

Differential Revision: D81518903

fbshipit-source-id: d3833e5d2a8279676e3674ae67c90f0b7d281a1e