Set outcome=Cancelled for IO_CLIENT_BROKEN_PIPE errors
Summary:
IO_CLIENT_BROKEN_PIPE means a file descriptor for stderr or stdout was closed before the command finished, typically because output was piped and the process it was pipe to died.

Changing the outcome to Cancelled causes any errors to be ignored. This could be a new outcome type but at a high level it makes sense to me to consider these commands cancelled and we shouldn't have more outcome values than necessary.

Motivation is that this error is the largest environment error by an order of magnitude and should be excluded if we want to count environment errors for reliability purposes, which we currently do as part of the overall (non-user) success rate.

Also renamed the exit code value to be more specific since this should only apply to this specific type of broken pipe.

Reviewed By: Will-MingLun-Li

Differential Revision: D84649995

fbshipit-source-id: 3b66a0477b4d2ddc65194f3c75079d0dc2e84071