cgroups: Introduce `BuckCgroupTree`
Summary:
In the next diff, I'm going to stop using systemd to spawn the forkserver. We'll need to create and know an alternative place in which to spawn it, so set that up.

All of the changes I'm making here will be gated behind enabling the action cgroup pool, that way we can evaluate them together and they don't risk breaking any existing thing.

I expect we will eventually want to unify this state with the memory tracker, but one thing at a time

Reviewed By: Will-MingLun-Li

Differential Revision: D84562795

fbshipit-source-id: 07095b7e756625bf5b6d2e821748b5c3d27cbcf0