cgroups: Kill forkserver cgroup pool tests
Summary:
Later in this stack, the cgroup pool is being moved out of the forkserver. As a result, we can't test the cgroup pool just in the forkserver anymore, which is a pretty fundamental assumption of these tests, meaning they're quite hard to fix.

I could attempt to completely re-write them, but that's pretty hard, and at the same time I don't know that I think they're adding that much value. We have some higher level e2e tests that assert that utilize the cgroup pool and assert that it works, I don't know that something like this which checks the exact reuse policy and cgroup structure is really adding all that much

I guess at the end of the day I'll leave it up to Nero5023 what to do about this, but I'd ideally prefer for that conversation to happen asynchronously (code that's uncommitted is not gone forever) so that it doesn't block moving forward with the cgroup pool experiment

Reviewed By: IanChilds

Differential Revision: D84572556

fbshipit-source-id: 564ddab2ad2980a909f9bb1b0271f7ab4d03d047