Add support for content-based promise artifacts
Summary:
Right now, we do not allow promise artifacts to resolve to an artifact that uses content-based paths. The reason for that is that `write` actions now need to know whether their inputs are content-based or not in order to know whether they need to depend directly upon them or not (if the input is content-based, the action needs to depend upon it directly in order to resolve the content-based path correctly). For promise artifacts we currently have no way of knowing if it is content-based or not, so right now we just disallow it.

This fixes that by adding a new param `content_based_artifact_promise_mappings` that should be used instead of `artifact_promise_mappings` is the artifact ends up resolving to a content-based path.

This diff is quite big, sorry about that but it didn't seem clear to me that there's anything that can be split out without making the diff incomplete.

Reviewed By: JakobDegen

Differential Revision: D83974779

fbshipit-source-id: 8dffac2c38196bac7cea9b1af70d8dcf37cb4f19