cgroups: Manage cgroup pool in the daemon
Summary:
Instead of holding the cgroup pool in the forkserver, hold it in the daemon. Then, on each request just send the forkserver the path to the cgroup and nothing else

For now, the pool is stored in the forkserver client in the daemon (that was the closest place to put it). Later, I'm going to push it further back into the memory tracker

This diff indirectly fixes a race condition in the implementation of the action cgroup pool. Previously, you might:

 1. In the forkserver, finish an action, send the command event for the action being finished, release the cgroup back to the pool
 2. In the forkserver, start a new action that acquire the same cgroup. Send the event that that action started (and the cgroup with which it started)
 3. In the daemon, receive the "command started" and "command ended" events out of order (since they're processed by different streams). When the action memory tracker receives them, its assertions that one cgroup is being used by only one action fire

With this change, the cgroup is not released until the action memory tracker has been updated, preventing the bug.

This prevention is not the most robust thing ever, we should remove the duplicate mutable state, but the rest will come later

Reviewed By: christolliday

Differential Revision: D84562783

fbshipit-source-id: a2b74910adfd2d6c4d481f2ec0ee2821051436ae