propagate what command execution metadata we have when cancelling actions
Summary:
Right now, if we cancel an action (because e.g. the whole test run timed out),
we do not log anything about its timing. Here is an example:

https://fburl.com/scuba/scuba_buck2_test_actions/vckmbwcw

This is annoying because it means that if some tests are so slow they always
cause the whole test run to timeout, we never learn about them (well, that's
one of two reasons for that, at least).

Reviewed By: JakobDegen

Differential Revision: D84694863

fbshipit-source-id: d3135d3a53dca9a321120c6df4bf47ef847eebf2