Fix error format parser to handle consecutive multi-line errors
Summary:
The error format parser had a bug that when there is no %Z (end of multiline), it will not count it as a vliad entry. This diff fix this issue.
When enter a new mutiline state or end of parsing, it will try to create a  valid entry if possible
It also aligns how vim errorfromat implement it

Reviewed By: Will-MingLun-Li

Differential Revision: D84668255

fbshipit-source-id: 4ca0269e10523ab74e408da15e385d3f4297eb60