Add I/O counts to InvocationRecord
Summary: For debugging materializer bottlenecks, I want to log number of I/O operations done in a build. We currently only log snapshots of I/O in flight, not overall through the build, so this diff adds setup to do that. We log count of all I/O operations done by daemon since beginning of the build in InvocationRecord sent.

Reviewed By: Will-MingLun-Li

Differential Revision: D84778403

fbshipit-source-id: 69a832d98d06c1599d3265cf7fdc1af3dea70491