Add is_eligible_for_dedupe to ArtifactGroup
Summary:
We want to add some logging that specifies whether an action is "eligible for dedupe" - that is, its inputs and outputs are declared in such a way that they can be shared across configurations.

This just adds a convenience method to `ArtifactGroup`. We say that for a given target platform, the `ArtifactGroup` is eligible for dedupe if (a) it itself is eligible for dedupe (most likely via being content-based), or (b) it is configured for a different platform.

(b) is intended to capture the case where we're configuring an execution dependency. Ideally, we'd actually check that it is equal to the execution platform, but `exec_cfg()` is not usually filled in, so I'm not sure how to get hold of the execution platform from the `ConfiguredTargetLabel`.

Reviewed By: scottcao

Differential Revision: D84604793

fbshipit-source-id: 92d23986d72bb576611a0396e0be423989d04a27