report queue time for cancelled RE actions
Summary:
Like it says in the title. However, keen observers will note that this does
nothing because we actually report the RE queue time via a completely
different field.

That is arguably a bug. The execution kind's supposed to represent what is
being executed, and the metadata is supposed to represent how it was executed,
but it looks like the RE queue time ended up in the wrong place.

It appears the only person we to blame for this is myself (D47754237); sadly
at this point it's not clear to me what I was thinking back then. I did try to refactor
this but there's a lot of push safety implications in doing so, so I didn't, so I did
fix this in the next diff.

Note: the implementation here will likely overcount a bit since we consider
actions are "queued" immediately upon being sent to RE, whereas in reality
there's a small time gap here.

Fixing this means either:

- Have two stages we track: the presentational stage and the real stage.
- Actually report a "re submitted" or some such stage but then we make our logs
  bigger.

Both of these seem rather unpleasant

Reviewed By: IanChilds

Differential Revision: D84907833

fbshipit-source-id: ddcf34c16c49ea15935b08d347edaa586ea33a78