Add a working .elp.tpml to the route of fbsource
Summary:
# Context
elp is the language server for Erlang developed at Meta. When an Erlang file is opened from VSCode (or elp ran from the terminal), it will look for the nearest `.elp.toml` in the ancestor directories, and use that for configuration.

# History
It used to be the case that elp would scan all directores from the one containing the `.elp.toml` file to find relevant apps, Erlang sources, etc. This made having a working `.elp.toml` at the root of fbsource infeasible, and we'd instead have one that would say "there are no Erlang files here". This meant that we'd have to include `.elp.toml` files where we did expect Erlang sources, namely, in `whatsapp/server` and `fbcode/buck2/prelude/erlang/common_test`.

Since then, elp has learnt to use `buck2` to find apps and sources, but we've never revisited the current setup

# Problem
With the current setup, we have some issues with files under `fbcode/buck2/prelude/erlang/common_test`:

  1. Ambiguity: they can be reached both from the waserver elp.toml  and from the buck2-prelude elp.toml. So when opening an Erlang file from the buck2-prelude, we can get one configuration or the other
  2. eqwalizer-specs: There's no eqwalizer-specs in the buck2-prelude configuration, so eqwalizer gives bogus and confusing errors.

# This diff
So let's unify things:
- We a top-level .elp.toml in waserver, that is a copy (symlink actually) of the one in waserver,
- We delete the one we have in the buck2 prelude, as it is now covered by the fbsource one
- We leave the one in waserver, in case there are things that depend on it being there

Reviewed By: michalmuskala

Differential Revision: D84927028

fbshipit-source-id: 29b73fd33b751c014bcc22857a94a09645dd95f6