Add new commandline argument
Summary:
Adding a new command line parameter for the new file that would contain the buck reports as they get streamed in.

We plan to generate a new report each time an output is generated (similar to the current behaviour of buck install).  Writing to a single file in json lines format seems to offer the best balance of simplicity of arguments and producer/consumer logic.

Reviewed By: Will-MingLun-Li

Differential Revision: D84511626

fbshipit-source-id: 1c9c587af041d33cc60dd0c8f6b8bfd8ce725d41