Ability to handle streaming build results
Summary:
Adding in the scaffolding to be able to handle the rx side of the build results. The function is getting bigger, but I'll leave it inline for the moment and extract the functionality at the top of the stack.

The actual code that will consume the build results comes in a later diff.

Reviewed By: Will-MingLun-Li

Differential Revision: D84056567

fbshipit-source-id: 5cea37a1c443041685f2e86f4f21c284811d25f4