Make BuildTargetResultBuilder::build callable multiple times
Summary:
The current implementation of BuildTargetResultBuilder::build consumes all the data in the BuildTargetResultBuilder so it can only be called once.

This diff changes it so the function clones the inputs so we can call it for both the intermediary build reports and the final one. I considered using two functions in order to keep the default behaviour identical, but maintaining two versions of the functionality seemed like a bigger downside that copying the data.

Reviewed By: Will-MingLun-Li

Differential Revision: D83833717

fbshipit-source-id: 06e76720aa4b6beea588ea20fc25bb97184011ad