Expose ApplePackageInfo containing information about the packaged output
Summary:
For the native mobile build rules that are beginning development, we need to take the output from an Apple Package and perform operations on it.

While we could pull from the `DefaultInfo` outputs, and read the file name, etc, this is less than ideal. Create and return `ApplePackageInfo` which contains basic information about the generated output and the handle to the artifact.

Reviewed By: milend

Differential Revision: D84946435

fbshipit-source-id: c7a33b623572d7a22c806d516fd2d60d31b7fe61