Add support for device type identifiers
Summary:
This stack fixes a problem where VS Code passes simulator names in to `buck test` as the device name. While the simulator name usually is the same as the device name, *this is not always true*. When the simulator name is not the device name, the resource broker will fail to setup a simulator.

This diff enables the device passed through the `apple.simulator_device` config to be a device name *or* **device type identifier**.

The next diff will update the VS Code codelens to pass the device type identifier instead of the simulator name.

Differential Revision: D85062379

fbshipit-source-id: d192be1f8e541e48a3adc6b673e3cb67f250ce54