Use argparse for test_binary argument parsing
Summary:
# Context
The `test_binary` launches a common-test runner on a separate runner and collects the results. As it was written before `argparse` made it to OTP in release 26, it uses a very simple positional parsing

# Problem
In order to make it easy to interface with the `edb` debugger, we would like to add an optional command line option to pass the `edb` code to inject on launch. But this becomes awkward with the current manual command line parsing

# This diff
We make the test_binary use `argparse`. For the arguments of the `list` and `run` we keep backwards compatibility until the tpx driver is updated.

Reviewed By: robertoaloi

Differential Revision: D85063388

fbshipit-source-id: d6741452eaeacefbe30109226c7197962a3d56b6