fix suggestion to include full path
This change returns a valid suggestion in the case of an invalid path to a BUCK file. This can come up from either making a typo in a target path or accidentially mismatching cases on a case-insensitive OS and then having a build run on a case-sensitive OS

Before, we matched on the closest parent directory then returned that as a suggested path, which can be confusing if `root//cat` is not a valid location:
```
dir `root//animals/pets/cats` does not exist. Did you mean `root//cat`?
```

After, we now return the full path to the matched directory allowing copy+paste of the suggestion to fix:
```
dir `root//animals/pets/cats` does not exist. Did you mean `root//animals/pets/cat`?
```

PS: I have no idea if the tests pass, the README says they can only run inside meta so this is just a guess