Add profile options to all commands
Summary:
This adds a common set of flags to all commands that support enabling
profiling based on regexes that match stringified forms of StarlarkEvalKind.

There's more documentation on the flags on the --profile-patterns flag in
CommonStarlarkOptions.

But basically:

--profile-patterns: accepts regex patterns of evaluations to enable profiling
--profile-patterns-mode: profiling mode to use
--profile-patterns-output: root path to write profiling data. every individual
evaluation gets its profiling data written to a separate file. we write to a
subdir of this with similar format to buck-out/log (i.e. datetime+uuid)

we additionally write an all_keys.list file listing all the evaluation names we profiled.

This works by (1) setting up profile configuration that StarlarkEvaluatorProvider uses to enable profiling as appropriate and then (2)
adding a ProfileEventListener that gets profile_collected calls as evaluations
are finished.

We already consistently use StarlarkEvaluatorProvider to get the enabling of
profiling.  The remaining diffs in the stack ensure that we consistently
finalize profiles for all evaluations.

Reviewed By: JakobDegen

Differential Revision: D64281578

fbshipit-source-id: b09824a94d1a596216fa9d30e761436492e63b39