Force starlark evaluations to report profiling results
Summary:
We now consistently enable profiling (and debugging) for starlark
evaluations, but it's possible for calls to not finalize the profiling data. In
that case, we don't get the profiling result or event.

This adds a simple BuckStarlarkModule wrapping starlark::Module and changes
apis to require that. Then, the only way to get that is by calling
`BuckStarlarkModule::with_profiling()` (or its async equivalent), which takes a closure that
is required to return a `Result<(ProfilingFinishedToken, R), E>`. I.e. it is required to
return a token that indicates it reported profiling results. The closure receives a
BuckStarlarkModuleProvider, which is easier to deal with than receiving a Module directly as
Module is not Send (and with Module there ends up being some additional
lifetime restrictions).

All callers are updated to use this.

Reviewed By: JakobDegen

Differential Revision: D84097281

fbshipit-source-id: 43415bc3ff95177af4b2f8184ce1a960541624b6