Prevent circular dependency if we have content-based paths and a pdb_artifact or split_debug_output
Summary:
We should only need to rewrite anything that has already been produced, and is not an output of the link action.

`pdb_artifact` and `split_debug_output` are produced by the link action, so writing them to the `separate_debug_info_args` doesn't make sense, since if they are content-based we won't have their final path, and if they are not content-based there is nothing more to do.

I believe it is the case that we never have a `pdb_artifact` or a `split_debug_output` when we have content-based paths, since otherwise we'd have a circular dependency here. But this protects against that changing in future and someone being really confused why that occurred.

It also means that we can enabling some checking to prevent that, in D84603090

Reviewed By: cjhopman

Differential Revision: D84698967

fbshipit-source-id: 97d31aede6aaab5f7cdcaaaf35748ac7648c059f