Add new buckconfig `buck2.disable_cell_segmentation`
ImportPath::new_with_build_file_cells takes a flag `cell_segmentation`
and suppresses the effect of build_file_cell if it is false.
We then plumb this boolean through buckconfig and dice to all the
locations it is needed to construct an ImportPath.

Ultimately if `buck2.disable_cell_segmentation` is true, then ImportPath
becomes a redundant wrapper around CellPath.