Use a buckconfig parser.cell_segmentation to enable/disable cell segmentation
ImportPath::new_with_build_file_cells takes a flag `cell_segmentation`
and suppresses the effect of build_file_cell if it is false.
We then plumb this boolean through buckconfig and dice to all the
locations it is needed to construct an ImportPath.

Ultimately if `parser.cell_segmentation` is false, then ImportPath
becomes a redundant wrapper around CellPath.