Move kotlin/cd/logger/scribe:scribe to prelude
Summary:
xpat and fbcode don't get along well and for that reason in D65832551 we decide to use org.apache.thrift from third-party.

It worked untill now that we need to use thrift for RE persistent worker and we ran into thrirf version mismatch.

This diff moves `kotlin/cd/logger/scribe:scribe` from xplat to fbcode so that we can move thrift dependency to the meta internal one.

Reviewed By: siaojiecai

Differential Revision: D84736217

fbshipit-source-id: e9837c00b2f1c3c0a2a833daefffe784f664fb20