Eqwalize epmd_manager
Summary:
# Context
The `epmd_manager` module deals with (optionally) starting and stopping an `epmd` on a random port, for exclusive use of a running suite

# Problem
The module is currently not eqwalized

# This diff
Eqwalize `epmd_manager`. After this, only 11/65 modules remain uneqwalized.

Interestingly, the `handle_info()` was broken, as we were matching an Erlang port against a TCP port, so all messages would go to the last clause, that would ignore them

Reviewed By: acw224

Differential Revision: D85247672

fbshipit-source-id: 473161aae1391cc4f56f3b3d18ba62c2c2f29743