Make it easier to add new CLI args on test_binary
Summary:
# Context
We recently started using `argparse` for handling command-line arguments to `test_binary`.

# Problem
We are currently deconstructing the args on the top-level handlers, which means that each time we need to add a new argument, we need to change a lot of code.

# This diff
Instead, delay the deconstruction until they are actually needed. We also introduce type aliases, which also simplifies adding new args

Reviewed By: michalmuskala

Differential Revision: D85247673

fbshipit-source-id: 945284078df71014839f5c1ddbe67d760372ae31