Add support for running tests under edb
Summary:
# Context
`edb` is the new Erlang debugger. In order to launch an Erlang program under `edb`, we need to inject some custom code when starting the debuggee node. This code is typically stored in some environment variable.

# Problem
The actual tests don't run on the same node that we launch to run tests. Instead the `test_binary` runs on the launched node, and it later spins up  a new node where the actual tests run. We need to inject the `edb` code in that node.

# This diff
We add an optional `--edb-code` to the `test_binary`, with the code that needs to be injected via `-eval`. To make integration as easy as possible, we make it so that an environment variable can be provided  as argument, instead of the actual code.

For example, in the non-tpx of the runner, one can do:

```
$ buck2 test my_SUITE -- --test-arg=--edb-code=EDB_CODE_TO_INJECT
```

Reviewed By: michalmuskala

Differential Revision: D85155456

fbshipit-source-id: dd1e6b5c19926b162a1c851a9f997b2e3c5192ba