remove property_test integration for ct_property_test
Summary:
I'm removing the support for ct_property_test in the buck2 rules, because the model it operates doesn't work well with buck2, and we see tests using it being flaky due to the compilation logic in ct_property_test.

Instead we can just use the `srcs` attribute on the erlang_tests macro to use buck2 for building the sources during the build process. This is faster and roboust.

Reviewed By: michalmuskala

Differential Revision: D85183016

fbshipit-source-id: a5f6668ab598eb33474d8fefeb5fc8c8e1dbfd78