Fix logic for formatting errors
Summary:
# Context
We currently extract the test result using a CT hook. In particular, in the `post_*` hooks we look at the result to decide if it is an error or not.

We then need to "unwrap" the error, and see if it comes with props that tells us how to format it properly

# Problem
The current logic for doing this was unclear and wrong. In particular, we would fail to apply custom formatting to certain erros in `init_per_*` functions.

The root of the problem was that CT seems to do several wrappings of the error terms:

- First there may usually a layer of `{skip, _}` or `{fail, _}`, that tells us things like "the test was skipped because the init_per_testcase` failed. We were already analyzing  this structure in the CT hooks, but we would then forward it to the formatting function, that would have to do seemingly random "unwrappings" of tuples.
- Second, for failures in `init_` or `end_` functions, there is an additional level of wrapping, one of the form: `{failed, Suite, Callback, ActualError}`. We were dealing with this in an ad-hoc way only for certain `end_per_testcase` cases, but not for `init_per_testcase`

# This diff
* When calling the formatting function, we omit the first level of unwrapping
* We introduce a formatting function that recieves also the context (suite, callback), so that it can pattern match to remove the second layer of wrapping
* Because of this, we remove the previous ad-hoc unwrappings

Reviewed By: acw224

Differential Revision: D85340961

fbshipit-source-id: aca87ab11c0e14e017bec5801b50ebd5bb2c142e