Move dep-file processing into 'filter'
Summary:
Instead of storing a `DirectorySelector` in `ConcreteDepFiles`, we store just the dep-file `String` representation, and then in `filter` we construct the `DirectorySelector` in `filter`.

This is a preparation diff for the next diff, where we are going to produce a different selector for each input directory, so we need to do the work in `filter` rather than when reading the dep-file.

Reviewed By: cbnilrem

Differential Revision: D84911236

fbshipit-source-id: a8a8b607d33ac6a47daec0122584c2da8984c77b