Add paths to the selector based upon content hashes from the input directory
Summary:
When we have content-based paths in a dep-file, the content-hash may
have changed since the dep-file was written in the last action, and we still
need to match on the new content-based path.

Therefore, when we find a path with a content hash in it, we look in the new
input directory to see which content hashes it may have been replaced by (or we
find the existing content hash), and we construct paths based on those hashes.

Reviewed By: JakobDegen

Differential Revision: D83846298

fbshipit-source-id: 230d8d9a5b9e68d348a43efc58ea0c84883d9b5c