Add expected_eligible_for_dedupe
Summary: This adds a mechanism to `run` actions to allow them to be specified as eligible for dedupe, which means that all of their outputs are content-based, and all of their inputs are eligible for dedupe. This can be used to prevent regressions to actions that have been migrated to be eligible for dedupe, since otherwise additional inputs/outputs can creep in that make them no longer eligible for dedupe.

Reviewed By: scottcao

Differential Revision: D85146295

fbshipit-source-id: ff1f80d3ceadb217b744c580a40279f8d43f5cb8