cgroups: Always use `CgroupResult`'s `memory.peak`
Summary:
We currently gather `memory.peak` in two places: miniperf and the action cgroup session. Only the second one should be used with the cgroup pool, so always rely on that.

In order to enforce that this is set, add some extra types that omit this field in places where it's not available

Reviewed By: christolliday

Differential Revision: D85419633

fbshipit-source-id: 0ebf03850a8f91ea30a2044f39d8148434399238