cgroups: Split `CgroupMinimal` out of `Cgroup`
Summary:
Shortly, I'm going to start opening some of the memory files when creating a `Cgroup`. About that though:

 1. That's a little wasteful in the forkserver where we don't need it
 2. We can't really do that when creating cgroups shortly after daemon startup, because can't enable the memory controller in some of the cgroups until after we move the daemon out of them

So split up the cgroup type into a more bare-bones one and a fuller one.

The fuller one says it has other files open - it doesn't yet, that comment is aspirational

Reviewed By: Will-MingLun-Li

Differential Revision: D85419643

fbshipit-source-id: 41575584bb969c7fc85e23f6c17b833964190738