buck2/rust: add [asm] subtarget for asm output
Summary:
This adds an asm subtarget like `[mir]` and `[llvm-ir]` subtargets. In fact so much like them I
consolidated all the logic so that it's not repeated.

This is supported for both rust_binary and rust_library targets. rust_binary still just produces the asm
output for the "main" crate, not the entire program.

Libraries: library:target[static][asm]
Binaries: binary:target[asm]

Reviewed By: dtolnay

Differential Revision: D85477639

fbshipit-source-id: 8426a4681d491d0c3137782975031f78b0cfc465