cgroups: Tell the forkserver fewer things it doesn't have to know about
Summary:
1. Get rid of `cgroup_command_id`, it's *almost* dead code
 2. Don't tell the forkserver at spawn time whether resource control is enabled or not. The only thing the forkserver needs to know is the cgroup path to use on each individual request

As a part of this, we can properly decouple the miniperf and cgroup handling in the forkserver.

Note that the `cgroup_command_id` being `None` was not *quite* dead code. It's the way we were preventing workers from being spawned in cgroups (was that obvious to you? It wasn't to me, I initially missed that until I spotted it later in the stack). We add in an explicit check on the command type to compensate for that

Reviewed By: christolliday

Differential Revision: D85419646

fbshipit-source-id: 14828fb6a977a1ce601648060781b4acd24ec908