cgroups: Spawn thread in forkserver
Summary:
See the comment for why

This is pretty ugly. If I had known about this issue earlier I would've designed this stack a little bit differently to potentially work around this; now though, that's a bit too much work.

There is a much better solution to this problem which is to make the forkserver responsible for actually carrying out the freezing operation instead of doing it from the daemon. That way it knows what's going on with the command when it does. I'll implement that later in the stack

Reviewed By: christolliday

Differential Revision: D85419638

fbshipit-source-id: f0b519d20e983cf431b297826cae0ab123be0bc0