take cxx_library link_group_info from impl params when provided
Summary:
The `link_group_info` cxx rule constructor param is used in `cxx_binary` to determine link group info:
https://www.internalfb.com/code/fbsource/[b79a6b1dae6f]/fbcode/buck2/prelude/cxx/cxx_executable.bzl?lines=306-308

This diff is adding similar support in `cxx_library`, with a fallback to the `get_link_group_info` function when not set. This allows for rules to have access to their link group info without having to recalculate it or pass it back.

Reviewed By: d16r

Differential Revision: D85446946

fbshipit-source-id: f32877ebb21d807bea4b73f3a37074b07e417468