Let test_binary choose the total timeout
Summary:
# Context
At the moment we impose an overall limit of 9m30s when a suite runs. This is handled in the `test_runner` module.

# Problem
In order to support debugging of tests via edb, we want to selectively turn off this timeout (as it doesn't make sense under debugging). But this means that we need to either add an extra arg to `test_runner:run_tests()` to let it know that we don't want a timeout, or we can  pass directly the timeout value to use. The latter is simpler.

# This diff
We refactor the code so that the decision of the timeout value is done in the `test_binary`. We'll handle the case of `edb` next in the stack

Reviewed By: TheGeorge

Differential Revision: D85449897

fbshipit-source-id: 7ddaaab0deb989bc9e30af6d1179756199d0169c