Deduplicate string/binary conversion functions
Summary:
# Context
The `unicode:characters_to_{binary,list}/1` functions pose a problem when eqwalizing as they can return error terms for unicode decoding errors that we never want to handle (let it crash), but break the calling contracts.

To workaround this, we've been defining several copies of the same conversion functions that just raise an error on decode errors, or even inlined this.

# Problem
We have now a lot of duplicated functions all over the place. Moreover, we already had a `common_util` module but was being used in few places.

# This diff
Move all definitions to `common_util` and use that instead, including for inlined definitions. We use an `import` statement to keep the calls shorter and also keep in line with previous usages of `common_util`.

Reviewed By: TD5

Differential Revision: D85564998

fbshipit-source-id: b5f409fcddb9a5d97c22a75deda79eeba398d583