Add missing specs to listing_interfacer
Summary:
# Context
The `listing_interfacer` uses `xmerl` to produce a listing file that TPX can consume

# Problem
While the module is currently eqwalized, it is missing some specs, so eqwalizer use `dynamic()` in those call, which prevents it from catching erros

# This diff
Add the missing specs. Because we rely on the `xmerl:simple_element()` type, that is not currently exported, we need to copy-paste it. Also, because the type doesn't guarantee that it accepts `binary()` as value of an attr, we perform a conversion to atom.

Reviewed By: michalmuskala

Differential Revision: D85594551

fbshipit-source-id: cdf6b9f90c1c9b016fcb177359cb188844d47a55