elp.bxl keep full include file name
Summary:
In the past we built a full include file path for a given target within ELP, based on its transitive dependencies and the includes from each.  And we only used the directories for this.

To reduce the processing in ELP we returned just the directories in the `includes` part per target.

We no longer do this, and are instead interested in the local includes for a given target only, so the performance optimisation is no longer needed.

There was a fidelity cost to this optimisation, where for any target selecting a subset of `.hrl` files from a given directory for a given target, this subset was lost as we flattened it to the directory as a whole.

This diff removes the flattening, returning the list of actual files included per target.  The update to ELP to match is in the following diff.

Reviewed By: TheGeorge

Differential Revision: D85573080

fbshipit-source-id: cb429af3d8152f03feb26a9f8c18cb8ef4d036dd