lifetimes: Put anon promises eval + dice access behind a trait
Summary:
Consider this code:

https://www.internalfb.com/code/fbsource/[021e4a27f584f28a1cceddd47a4be7b6c31f517e]/fbcode/buck2/app/buck2_build_api/src/interpreter/rule_defs/context.rs?lines=85-104

There's something very wrong here: This is going to run the anon promise mappings, and those mappings have access to a `&mut DiceComputation` through the bxl context. But somehow, we've also passed the same `&mut DiceComputations` into the function.

The root cause of this is some unsoundness in starlark that I'm not going to get into right now (when a later diff talks about "missing branding," it's that).

Begin to fix this by putting access to dice and the evaluator behind a trait

Reviewed By: cjhopman

Differential Revision: D85599714

fbshipit-source-id: a40dae95be19573ada63bae2bfdd057d9ce07f2f