enables killall to work when the binary has been deleted
This change makes `buck2 killall` work on linux when the underlying
binary has been deleted.

Currently on linux if the buck2 binary has been deleted, `buck2 killall`
returns "No buck2 processes found" and fails to kill the buck2 process.
This is because we check the symlinked path in `/proc/<pid>/exec` for
the string 'buck2'. Linux appends the string ` (deleted)` to the
symlinked path in the case the binary was deleted, causing the
`is_buck2_exe` check to fail