lifetimes: Require evaluator in `via_dice` in bxl
Summary:
Let me briefly use this to explain what I'm doing: In an effort to fix the unsoundness identified some diffs ago, and also in an attempt to clean some other stuff up, I'm going to move the access to the `&mut DiceComputations` off of the `BxlContext` and onto the `Evaluator`.

This diff unblocks that by making the evaluator available in the right place

There are two pieces of code that must be exempted from this, they're the ones that cannot be made to work without fixing the unsoundness. A subsequent diff will do that

Reviewed By: Nero5023

Differential Revision: D85599733

fbshipit-source-id: 1e74d8c08959cd1d4d27bbc49d34417d7865e20a