lifetimes: Take `DiceComputations` out of the starlark heap
Summary:
Alright, so this is a complicated diff, and I apologize for that, but unfortunately it's not clear how to split it up.

The changes can be understood as the following, in about this order:

 1. Remove `async_ctx` and dice access from `BxlContext`
 2. Remove the `Rc<RefCell<_>>` wrapping the `dyn BxlDiceContext` in the eval context. We now only keep dice in the evaluator. Store dice in the `extra_mut` instead of the `extra` as a part of this
 3. Adjust the dice access API on the `RunAnonPromisesAccessor` to look more like the dice access API in bxl, specifically so that it has the `for<'d> FnOnce(&'s mut DiceComputations<'d>)` in the signature. This is a bit boilerplatey but callsite adjustments to support it are mostly trivial
 4. In the `BxlAnonPromisesAccessor`, store only the evaluator, not a `&mut DiceComputations`, and use the standard dice access API on the evaluator to access dice from the aforementioned method.
 5. This in turn removes the need to use the `BxlContext` to access dice for the `BxlAnonPromisesAccessor`, which a) fixes the unsound thing from some diffs ago, and b) removes the last use of `async_ctx` on `BxlContext` completing the circular dependency on step 1

I'll call out one important thing here: In order to make step 3 work, we need to add a method to `dyn BxlDiceContext` that is like `via_dice` but does not internally await the future, instead just returning it to be awaited by the caller. This is needed because the other implementation of `RunAnonPromisesAccessor` for non-bxl doesn't have to do any of this hoop jumping and wants to just look like normal async code with the caller, `run_anon_promises` doing the awaiting. However, this means that we then have to explicitly `.block_on` the resulting future, essentially bypassing the safety that the safe bxl computations provides. Chris' TODO is about that issue being somewhat pre-existing though, and I may try and clean this up in a following diff

Reviewed By: Nero5023

Differential Revision: D85599717

fbshipit-source-id: 7f2b3f35205f72dddfb85e8d59c534e200849709