Add helper for local linker arg generation
Summary:
We create a closure that encapsulates some of the logic used to craft local link invocations. This closure takes as a parameter a boolean that controls if linker output artifacts are added to the created cmd_args.

In a future diff, we will invoke this helper again, but opt out from adding linker outputs, to instead create a second linker invocation that produces an interface for the link unit rather than any of the conventional linker outputs.

Reviewed By: rmaz

Differential Revision: D85703562

fbshipit-source-id: 5d09e343ea2aae31affe1b2ca07487b2bebd0430