Remove the execute bit from python cxx scripts (#1134)
Summary:
These scripts are not being executed directly, so the execute bit is never really used.  The reason why it is beneficial to remove the execute bit is that when compiling for Linux from Windows, the python scripts are sent without the execute bit being set.  This means that the input files will differ between the scenario where a Windows machine is building for Linux versus the scenario where a Linux machine is building for Linux.  This ruins the ability for anything to be cached.  By simply removing the execute bit here, this helps the input files become the same whether the client is on Windows or Linux.

Pull Request resolved: https://github.com/facebook/buck2/pull/1134

Reviewed By: dtolnay

Differential Revision: D85894586

fbshipit-source-id: 5b42b834cdcc174f10cf2fc5fc3edde493adff8c