use RE Client QueueStats for all RE actions
Summary:
I added this in D84907833 to capture it for cancelled RE actions, but then
we discovered that the whole codepath is completely busted for not-cancelled
RE actions anyways:

https://fb.workplace.com/groups/1734123223981259/posts/1872605076799739/?comment_id=1872763750117205&reply_comment_id=1873335063393407

So, let's just use this everywhere. There actually isn't an exact way to
capture the right number as it turns out: D85943761.

Reviewed By: JakobDegen

Differential Revision: D85950588

fbshipit-source-id: 20906a3608ae02a9c47ceef8ac466f542d346cec