Always attach StarlarkContext during conversion to buck2_error::Error
1. Always attach
2. Blindly push a StarlarkContext context value in error_with_starlark_context
   instead of injecting it as close as possible to the root.

Previously it only got added for starlark ErrorKind::Native, but we need
it for every kind of evaluation error. We need to give the user a span
for every kind of thing that goes wrong in starlark. This is for LSP.

This also changes the effect of skip_stacktrace to operate on the StarlarkContext,
which is necessary to make it still make sense.
I checked this works for the one place we use skip_stacktrace=True
i.e. bxl's fail_no_stacktrace()

LSP needs the span. We need to always attach StarlarkContext so the span
is not lost. But we need to do more than that -- we need to delete any
code that throws away StarlarkContext.

In inject_starlark-context, we were previously just formatting
StarlarkContext & overwriting the root's description with that. This is
unacceptable, as it loses the span by rendering it into a string. So
now, we just blindly add StarlarkContext on top.

We rejig the error_with_starlark_context tests to just make assertions about
non-repetition, instead of the pop/replace with StarlarkContext
behaviour.