Always attach StarlarkContext
... when converting starlark errors to buck2_error::Error. Previously it
only got added for starlark ErrorKind::Native, but we need it for every
kind of evaluation error. We need to give the user a span for every kind
of thing that goes wrong in starlark.

This also changes the effect of skip_stacktrace to operate on the StarlarkContext,
which is necessary to make it still make sense.
I checked this works for the one place we use skip_stacktrace=True
i.e. bxl's fail_no_stacktrace()