Fix failing tests
Summary:
In D85674569, we attempted to mitigate S576379, which concerns increased OOMs in Roboelectric test jobs. The goal was to assess whether disabling KSIC would help reduce the OOM ratio.

However, the initial solution caused some tests to fail (T243435873), particularly those involving calls from the prelude cell, indicating that the approach was not universally effective.

This diff addresses the issue by relocating the selection logic outside of the prelude.

Reviewed By: hick209, siaojiecai

Differential Revision: D85952516

fbshipit-source-id: 9cd7ae960676469eb0d8550c0f734d4107eda9cb