Make StarlarkContext replace the root error, at format time
LSP needs the span. We need to always attach StarlarkContext so the span
is not lost. But we need to do more than that -- we need to delete any
code that throws away StarlarkContext.

In inject_starlark-context, we were previously just formatting
StarlarkContext & overwriting the root's description with that. This is
unacceptable, as it loses the span by rendering it into a string. So
now, such a StarlarkContext sits right above the root and says "i have
the same error message as root, but with a span, so render me instead."

An alternative would be to make ErrorRoot span-aware and then handle
both StarlarkContext and ErrorRoot when doing the callstack collapsing
etc.

We rejig the inject_starlark_context tests to just make assertions about
non-repetition, instead of the pop/replace with StarlarkContext
behaviour.