Make StarlarkContext replace the root error at format time only
LSP needs the span. We basically need to always attach StarlarkContext so the
span is not lost.

In error_with_starlark_context, we were previously
just formatting StarlarkContext & overwriting the root's description with that. This is unacceptable. So now,
such a StarlarkContext sits right above the root and says
"i have the same error message as root, but with a span,
  so render me instead."

An alternative would be to make ErrorRoot span-aware and then
handle both StarlarkContext and ErrorRoot when
doing the callstack collapsing etc.

We rejig the inject_starlark_context tests to just
make assertions about non-repetition, instead of
the pop/replace with StarlarkContext behaviour.