Don't symlink native libs in editable install
Summary:
We rely on `RPATH`s set on the native libs and extensions built and
deployed in the wheel.  However, in editable mode, we assemble a
symlink tree, which contains symlinks to these native libs, which breaks
these `RPATH`s, as they resolve relative their realpath.

This diff fixes to copy native libs into the editable tree instead of
symlinking as, unlike Python sources, there isn't an expectation that
we can modify and run in-place, w/o a build.

Reviewed By: christycylee

Differential Revision: D85963325

fbshipit-source-id: 698c9615ff8523de0e52fa25affca98c112e3026