cgroups: Implement kill-and-retry
Summary:
Like it says in the title. This is pretty straightforward now. We set up some channels associated with the `ActionCgroupSession`, and when we want to kill something, we notify the local executor via that channel, it cancels the command, waits on another channel for the retry notification, and then restarts it.

In the resource controller we treat these kill-and-retrys as being like a freeze, but implemented slightly differently. In some future diff I'll replace the "freeze" terminology with "suspend," to encompass both the "cgroup freeze" and "kill and retry" implementations of the suspend logic.

There's a whole bunch of work left to do; there is telemetry (on account of us reusing the existing freeze stuff) but not enough, the freeze and unfreeze algorithms are not really suited to this thing at all, etc. All follow-ups.

Reviewed By: Will-MingLun-Li

Differential Revision: D85419635

fbshipit-source-id: 877ad8467a1394d53c132b306e2d22abdc819942