Use label to materialize secondary native debuginfo
Summary: Add a label to indicate a target requires secondary native debuginfo to be materialized. We're going with a label approach instead of reading a conditional modifier constraint because BXL is only able to read CLI-based modifiers at the moment. Another part of the process will assign this label based on a conditional modifier constraint, though.

Reviewed By: patskovn

Differential Revision: D84785332

fbshipit-source-id: 4fd4f6397b96c37222ed2fd4cb813e2479fb8356