cgroups: Rewrite `use_some_memory` tool
Summary:
I'm going to guess no one had looked very closely at what this thing was doing.

The first thing to note is that some of the logic was just completely weird - I don't know why it was trying to "detect memory pressure," doing so by timing how long allocations take, and then doing that wrong.

But also the more general strategy of attempting to do any of this via memory locking is pretty sketch - that prevents swapping, which is going to cause the exact opposite of the pressure that we want to create (it'll create pressure in other places, but that's not really right).

This is simpler. Also, I added in a component that busy-loops and detects when freezes happen. Next diff uses that to further improve the tests

Reviewed By: cbnilrem

Differential Revision: D86075945

fbshipit-source-id: a29c4bbb14d83bb99ecca3ec135b6fb07095bf02