Add `.package` api for ProvidersLabel
Summary:
We change the label of ConstraintValue from TargetLabel to ProvidersLabel. To make sure backwards compatibility, make sure most apis of TargetLabel be same as ProvidersLabel.

Right now after this change only `with_sub_target` api of TargetLabel is not available in  ProvidersLabel. This would not be used in the current usage of constraint.

Context: https://fb.workplace.com/groups/buck2users/posts/4138120303110986

Reviewed By: christolliday

Differential Revision: D86538316

fbshipit-source-id: 979451d992d20bc6f33f5a0b02dd0ef2d4b21924