Move iOS simulator setup logic to build defs
Summary:
Currently, if an Apple test is executing "locally" (meaning not in RE) it's assumed that a local simulator is required for any device that needs one. This is no longer true if a test command is initiated in an OD, but actually executes on the user's machine (which is "not RE").

This diff moves the logic `tpx` uses to decide if a simulator is needed to the build defs.

Differential Revision: D86558105

fbshipit-source-id: 3e4a84f9590fa082afb8dcb056211de8c2573408