Rename data.proto stdout/stderr properties
Summary:
The CI log parsing tool that I'm writing in C++ takes a dependency on these protos, and (as I've discovered) needs to work on Windows. Unfortunately ` fbcode//mode/opt-win` has the side effect of `stdin`/`stdout`/`stderr` being declared as macros and that conflicts with the properties generated by the proto compiler.

First I tried doing some `undef`ing the macros before including the generated headers, but that doesn't work because the `.cc` files include their own headers (which still have the problem). I also tried adding a build rule to include the `undef`s at the beginning of each generated file but I couldn't get that to work cross-plat (and that seems fragile anyway). Just renaming the field seems fine since we're not changing the tag numbers or doing anything with reflection AFAICT. I've added a `serde` annotation to each field so that if anything is relying on the old JSON name it shouldn't break.

Reviewed By: JakobDegen

Differential Revision: D86560858

fbshipit-source-id: 4d1732dd41b4448564bd22d42cce1caa029d6ab4